requirements.txt
requests>=2.25.1
scikit-image>=0.18.1
viaa-chassis>=0.2.0rc1
jpylyzer>=2.0.0
PyExifTool>=0.4.9
Pillow>=9.5.0
Wand>=0.6.13
retry>=0.9.2

#######################
# Watcher requirements
#######################
inotify>=0.2.10
pygfried>=0.7.0

#######################
# Consumer requirements 
#######################
pika>=1.3.2
mediahaven>=0.7.0

